Create the UI for the calculator
Most of the html elements on the page use flex displays and border-box
box sizing. Border-box sizing was chosen because a lot of pixel-perfect
calculations had to be made when trying to align buttons inside the
body of the calculator.

The widths and heights of each button were also stored in css variables
so they can be adjusted dynamically based on how much gap I wanted to
put in between each button.