Tweak the style of the calculator
Make the calculator look prettier in css. Javascript was also used to
change the color of the button backgrounds when the mouse is
interacting with them.

There was also a bug where inputting a 0 after a "0." on the display
would clear the display back to "0". This bug was fixed so it is
possible to type a value like "0.005".